created table component
