+++ src/app/modules/report/block-list/block-list.component.ts
@@ -3,10 +3,8 @@ import { MatTableDataSource, MatPaginator } from '@angular/material';
 import { UtilityService } from 'shikshalokam';
 import { ReportService } from '../report-service/report.service';
 import { ActivatedRoute, Router } from '@angular/router';
-import { GlobalConfig } from 'src/app/global-config';
 import { SelectionModel } from '@angular/cdk/collections';
 import { MatBottomSheet } from '@angular/material';
-import { ActionSheetComponent } from '../action-sheet/action-sheet.component';
 
 @Component({
   selector: 'app-block-list',
@@ -15,8 +13,6 @@ import { ActionSheetComponent } from '../action-sheet/action-sheet.component';
 })
 export class BlockListComponent implements OnInit {
   object = Object;
-  displayedColumns: string[] = ['select', 'name', 'city', 'actions'];
-  columnsForBlockTable: string[] = ["labels", "action"];
   expandedElement;
   blocks;
   dataSource;
@@ -26,12 +22,9 @@ export class BlockListComponent implements OnInit {
   blockId;
   error: any;
   selection;
-  numSelected;
-  numRows;
   arr = [];
   enableMultiSchool: boolean = false;
   blockListDataSource;
-  links = {};
 
 
 
@@ -46,52 +39,7 @@ export class BlockListComponent implements OnInit {
   constructor(private bottomSheet: MatBottomSheet, private route: ActivatedRoute, private reportService: ReportService, private utility: UtilityService, private router: Router) {
     this.route.queryParams.subscribe(params => {
       this.programId = params["ProgramId"];
-      this.links = {
-        multiEntity: [
-          {
-            label: "Drilldown Report",
-            link: "/report/multiple-entity-drilldown-report/",
-            params: "",
-            queryParams: {
-              ProgramId: this.programId,
-              school: "",
-              blockName:""
-            }
-          },
-          {
-            label: "Highlevel Report",
-            link: "/report/multiple-entity-report",
-            params: "",
-            queryParams: {
-              ProgramId: this.programId,
-              school: "",
-              blockName:""
-            }
-          }
-        ],
-        singleEntity: [
-          {
-            label: "Drilldown Report",
-            link: "/report/entity-report/",
-            params: "",
-            queryParams: {
-              ProgramId: this.programId,
-              // school:""
-            }
-          },
-          {
-            label: "Highlevel Report",
-            link: "/report/highlevel-entity-report/",
-            params: "",
-            queryParams: {
-              ProgramId: this.programId,
-              // school:""
-            }
-          }
-        ]
-      };
     });
-    // this.blockClick(this.blockId);
   }
 
   ngOnInit() {
@@ -109,17 +57,14 @@ export class BlockListComponent implements OnInit {
       },
         (error) => {
           this.error = error;
-          this.snackBar.open(GlobalConfig.errorMessage, "OK", { duration: 9000 })
+          // this.snackBar.open(GlobalConfig.errorMessage, "OK", { duration: 9000 })
           this.utility.loaderHide();
           ;
         }
       );
   }
 
-  toggleRow(row) {
-    this.selection.toggle(row);
-    this.enableMultiSchool = this.selection.selected.length > 1 ? true : false;
-  }
+  
 
   getSchoolList(id) {
     // this.utility.loaderShow();
@@ -130,7 +75,7 @@ export class BlockListComponent implements OnInit {
         this.paginator.pageSize = 5;
         this.paginator.pageIndex = 0;
         this.paginator.length = data['result']['schools'].length;
-        // console.log(this.paginator)
+        console.log(this.paginator)
         this.dataSource.paginator = this.paginator;
         this.selection = new SelectionModel(true, []);
       },
@@ -142,58 +87,6 @@ export class BlockListComponent implements OnInit {
         }
       );
 
-  }
-
-  isAllSelected() {
-    const numSelected = this.selection ? this.selection.selected.length : 0;
-    const numRows = this.dataSource.data.length;
-    return numSelected === numRows;
-  }
-
-  /** Selects all rows if they are not all selected; otherwise clear selection. */
-  masterToggle() {
-    this.isAllSelected() ?
-      this.selection.clear() :
-      this.dataSource.data.forEach(row => this.selection.select(row));
-    this.enableMultiSchool = this.selection.selected.length > 1 ? true : false;
-
-  }
-
-  checkboxLabel(row?): string {
-    if (!row) {
-      return `${this.isAllSelected() ? 'select' : 'deselect'} all`;
-    }
-    return `${this.selection.isSelected(row) ? 'deselect' : 'select'} row ${row.position + 1}`;
-  }
-
-
-  goToSingilEntityReport(id) {
-    console.log("clicked", id);
-    this.router.navigate(['/report/block-list/'], { queryParams: { ProgramId: this.programId, Id: id } });
-  }
-
-  getAction(actionFor, blockName, schoolId?: any) {
-    if (actionFor === 'multiEntity') {
-      const schoolArray = []
-      for (const item of this.selection.selected) {
-        schoolArray.push(item._id);
-      }
-      for (const link of this.links[actionFor]) {
-        link.queryParams.school = Object.assign([], schoolArray);
-        link.queryParams.blockName = blockName;
-      }
-    } else {
-      for (const link of this.links[actionFor]) {
-        link.params = schoolId;
-      }
-    }
-
-    console.log(this.links[actionFor]);
-    this.bottomSheet.open(ActionSheetComponent, { data: this.links[actionFor] })
-  }
-
-  applyFilter(filterValue: string) {
-    this.dataSource.filter = filterValue.trim().toLowerCase();
   }
 
 }
