+++ src/app/modules/report/table/table.component.scss
@@ -1,0 +1,4 @@
+@import '../../../../assets/scss/varibles.scss';
+.width{
+    width:100%;
+}
